+++ app/controllers/proposal_controller.py
@@ -9,6 +9,8 @@ from sqlalchemy.exc import IntegrityError
 from app.configs.database import db
 from http import HTTPStatus
 
+from app.models.provider_model import Provider
+
 
 session: Session = db.session
 
@@ -20,65 +22,72 @@ def get_proposals():
 
 
 @jwt_required()
-def get_proposal_accepted():
+def get_accepted_proposals():
     try:
         current_user = get_jwt_identity()
 
         if current_user["type"] != "provider":
-            return {"error": "acess denied"}, HTTPStatus.UNAUTHORIZED
+            raise Unauthorized
 
         call_list = Call.query.all()
-        proposals_list = Proposal.query.filter_by(id=current_user["id"])
 
         final_list = []
 
         for call in call_list:
-            if call.selected_proposal != None:
-                for proposal in proposals_list:
-                    if call.selected_proposal == proposal.id:
-                        final_list.append({"call": call, "proposal": proposal})
+            if (
+                call.selected_proposal != None
+                and call.selected_proposal.provider_id == current_user["id"]
+            ):
+                final_list.append({"call": call, "proposal": call.selected_proposal})
         return jsonify(final_list), HTTPStatus.OK
+
     except BadRequest as e:
         return e.description, HTTPStatus.BAD_REQUEST
+
+    except Unauthorized:
+        return {"error": "access denied"}, HTTPStatus.BAD_REQUEST
 
 
 @jwt_required()
 def create_proposal():
-
     current_user = get_jwt_identity()
-
-    if current_user["type"] != "provider":
-        return {"error": "access denied"}, HTTPStatus.BAD_REQUEST
-
-    data = request.get_json()
-    data["provider_id"] = current_user["id"]
-
-    default_keys = ["price", "description", "call_id"]
-
-    for key in default_keys:
-        if key not in data.keys():
-            return {
-                "error": f"Incomplete request, check {key} field"
-            }, HTTPStatus.BAD_REQUEST
-
-    for key in data.keys():
-        if key not in default_keys:
-            return {
-                "error": f"Incomplete request, check {key} field"
-            }, HTTPStatus.BAD_REQUEST
-
     try:
-        proposal = Proposal(**data)
+        if current_user["type"] != "provider":
+            raise Unauthorized
+
+        data = request.get_json()
+
+        provider_id = current_user["id"]
+        call_id = data.pop("call_id")
+
+        provider: Provider = session.query(Provider).get_or_404(
+            provider_id, description={"error": "provider doesn't exist"}
+        )
+
+        call: Call = session.query(Call).get_or_404(
+            call_id, description={"error": "call doesn't exist"}
+        )
+
+        valid_data = Proposal.check_fields(data)
+
+        proposal = Proposal(**valid_data)
+
+        provider.proposals.append(proposal)
+        call.proposals.append(proposal)
 
         session.add(proposal)
         session.commit()
 
+        return jsonify(proposal), HTTPStatus.CREATED
+
     except BadRequest as e:
         return e.description, HTTPStatus.BAD_REQUEST
+
     except IntegrityError:
         return {"error": "Proposal already registred"}, HTTPStatus.CONFLICT
 
-    return jsonify(proposal), HTTPStatus.CREATED
+    except Unauthorized:
+        return {"error": "access denied"}, HTTPStatus.BAD_REQUEST
 
 
 @jwt_required()
@@ -86,17 +95,19 @@ def update_proposal(proposal_id):
 
     current_user = get_jwt_identity()
 
-    if current_user["type"] != "provider":
-        return {"error": "access denied"}, HTTPStatus.BAD_REQUEST
-
     try:
+
+        if current_user["type"] != "provider":
+            raise Unauthorized
+
         data = request.get_json()
 
-        proposal: Proposal = Proposal.query.get(proposal_id)
+        proposal: Proposal = Proposal.query.get_or_404(proposal_id)
 
-        allowed_columns = ["price", "description"]
+        if proposal.provider_id != current_user["id"]:
+            raise Unauthorized
 
-        valid_data = {item: data[item] for item in data if item in allowed_columns}
+        valid_data = Proposal.check_data_for_update(data)
 
         for key, value in valid_data.items():
             setattr(proposal, key, value)
@@ -108,8 +119,12 @@ def update_proposal(proposal_id):
 
     except BadRequest as e:
         return e.description, HTTPStatus.BAD_REQUEST
+
     except NotFound:
-        return {"error": "no data found"}, HTTPStatus.NOT_FOUND
+        return {"error": f"Proposal {proposal_id} do not found"}, HTTPStatus.NOT_FOUND
+
+    except Unauthorized:
+        return {"error": "access denied"}, HTTPStatus.BAD_REQUEST
 
 
 @jwt_required()
@@ -117,7 +132,11 @@ def delete_proposal(proposal_id):
     current_user = get_jwt_identity()
 
     try:
-        proposal = Proposal.query.get(proposal_id)
+
+        proposal: Proposal = Proposal.query.get_or_404(proposal_id)
+
+        if proposal.provider_id != current_user["id"]:
+            raise Unauthorized
 
         session.delete(proposal)
         session.commit()
@@ -125,7 +144,6 @@ def delete_proposal(proposal_id):
         return "", HTTPStatus.OK
     except BadRequest as e:
         return e.description, HTTPStatus.BAD_REQUEST
-    except UnmappedInstanceError:
-        return {
-            "error": f"Proposal {proposal['id']} do not found"
-        }, HTTPStatus.NOT_FOUND
+
+    except NotFound:
+        return {"error": f"Proposal {proposal_id} do not found"}, HTTPStatus.NOT_FOUND
